backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.6
djangorestframework==3.14.0
h11==0.14.0
idna==3.4
outcome==1.3.0
packaging==23.2
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.11.1
urllib3==2.0.7
webdriver-manager==4.0.1
wsproto==1.2.0
